Array2Pointer: Rewrite GEP handling
A single getelementptr (GEP) instruction can contain multiple indices.
The first must be interpreted in terms of the size of the pointer value
passed as the second operand to GEP. If we change the type of the
pointer value we must update this indices. Currently the GEP is
unrolled and the first index is updated.